Move `HTTPClientStub` into its own file
